Optimise part 2 with that initial plan
This commit improves part to run in close to 1 ms whereas before it
took around 7 minutes on my machine. The optmisation is no structural
exploit, but a simple heuristic optmisation that I wanted to implement
in the first place.

Basically, instead of brute forcing all seed ranges by mapping all
possible seeds one-by-one, we could map the seed ranges themselves.
The problem with this was that there might be multiple mappings for
a single input seed range. This just means that a seed range may split
up into different sub-ranges that can be mapped fully together. (I am
ignoring the part the sub-ranges may even need to split further, as
it can become a recursive explanation, but this case is treated in code,
by splitting valid ranges until no more need to split.)

The main issue was implementing this splitting requires care, which
was annoying for a day5 challenge. For example, the last case in
`Range.split(...)` was not there and it kept crashing on me (rightfully).